% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal.positions.R
\name{animal.positions}
\alias{animal.positions}
\title{Function that outputs a scatterplot of all observations of one ID (animal)
with longitude as the x-axis and latitude as the y-axis and the colour of
each point represents HPE.}
\arguments{
\item{animal_tags}{Table of animal data}

\item{ani_hpe_col_name}{String name of column that has HPE values}

\item{lat_col_name}{String name of column that has latitude of observations}

\item{long_col_name}{String name of column that has longitude of observations}

\item{num_bins}{Integer number of bins that user wants HPE to be split into}

\item{ID}{Integer of animal ID}
}
\description{
Function that outputs a scatterplot of all observations of one ID (animal)
with longitude as the x-axis and latitude as the y-axis and the colour of
each point represents HPE.
}
\examples{
anipos_id <- c(15048, 18032, 15100, 15075, 15086)
for(x in anipos_id) {
  animal.positions(animal_tags=animal_tag_data, ani_hpe_col_name="HPE",
                   lat_col_name="Latitude", long_col_name="Longitude",
                   num_bins=6, ID=x)
}

}
