% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve.perclost.R
\name{curve.perclost}
\alias{curve.perclost}
\title{Outputs a plot that shows the percentage of data remaining after filtering
by each HPE value.}
\usage{
curve.perclost(sync_tags, hpe_col_name, hpe_val)
}
\arguments{
\item{sync_tags}{Table of data of sync tags}

\item{hpe_col_name}{String name of column in sync_tags that has HPE values}

\item{hpe_val}{Integer value for max HPE value user wants to display the graph to (with the max being the largest HPE value rounded to the nearest integer using base R round, IEC 60559 standard)}
}
\description{
Outputs a plot that shows the percentage of data remaining after filtering
by each HPE value.
}
\examples{
# Load data
sync_tag_data <- read.table(file = './extdata/dummy_sync.csv', header=TRUE, sep=",")
# Example
curve.perclost(sync_tags=sync_tag_data, hpe_col_name="HPE", hpe_val=20)

}
