% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_perclost.R
\name{curve_perclost}
\alias{curve_perclost}
\title{Outputs a plot that shows the percentage of data remaining after filtering
by each HPE value.}
\usage{
curve_perclost(sync_tags, hpe_col_name, hpem_col_name, hpe_val)
}
\arguments{
\item{sync_tags}{Table of data of sync tags}

\item{hpe_col_name}{String name of column in sync_tags that has HPE values}

\item{hpem_col_name}{String name of column in sync_tags that has HPEm values}

\item{hpe_val}{Integer value for max HPE value user wants to display the graph to (with the max being the largest HPE value rounded to the nearest integer using base R round, IEC 60559 standard)}
}
\description{
Outputs a plot that shows the percentage of data remaining after filtering
by each HPE value.
}
\examples{
# Load data
sample_file <- system.file("extdata", "dummy_sync.csv", package = "filterTelemetry")
sync_tag_data <- read.table(file = sample_file, header=TRUE, sep=",")
# Example
curve_perclost(sync_tags=sync_tag_data, hpe_col_name="HPE", hpem_col_name="HPEm", hpe_val=20)

}
