% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.numdet.R
\name{filter.numdet}
\alias{filter.numdet}
\title{Finds a list of which values in a column occur more than a specified number of times in a dataset}
\usage{
filter.numdet(column.name, num.det)
}
\arguments{
\item{column.name}{Name of column to count}

\item{num.det}{Minimum number of times occurred}
}
\value{
A list of numerics contained in the column that occur more than num.det times
}
\description{
Finds a list of which values in a column occur more than a specified number of times in a dataset
}
\examples{
anitags_200 <- filter.numdet(animal_tag_data$Id, 200)

}
