% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpe.quantiles.R
\name{hpe.quantiles}
\alias{hpe.quantiles}
\title{Get HPE quantiles}
\usage{
hpe.quantiles(sync_tags, hpe_col_name, quantiles = NULL)
}
\arguments{
\item{sync_tags}{Table of data of sync tags}

\item{hpe_col_name}{String name of column in sync_tags that has HPE values}
}
\description{
Get HPE quantiles
}
\examples{
num_obs_per_tag_lob <- num.obs.per.tag(animal_merged, "Id", graph=TRUE,
sp="Homarus americanus", sp_col_name="Sp")

}
