% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpe.stats.R
\name{hpe.stats}
\alias{hpe.stats}
\title{Get number of observations per tag when given all data or data is sorted
by species, sex, grouping, or a combination, can output graph. Grouping can
be used for a variety of different separations of data such as treatment.}
\usage{
hpe.stats(sync_tags, hpe_col_name)
}
\arguments{
\item{sync_tags}{Table of data}

\item{hpe_col_name}{String name of column that has tag IDs}
}
\value{
Vector of number of total number of observations and total number of tags (for given parameters) and graph if graph=TRUE
}
\description{
Get number of observations per tag when given all data or data is sorted
by species, sex, grouping, or a combination, can output graph. Grouping can
be used for a variety of different separations of data such as treatment.
}
\examples{
num_obs_per_tag_lob <- num.obs.per.tag(animal_merged, "Id", graph=TRUE,
sp="Homarus americanus", sp_col_name="Sp")

}
