% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpe_linear_regression.R
\name{hpe_linear_regression}
\alias{hpe_linear_regression}
\title{Calculate linear regression given HPE level or percentage of data.
Accepts only one of hpem_max or percentage.}
\usage{
hpe_linear_regression(
  sync_tags,
  hpe_col_name,
  hpem_col_name,
  id_col_name,
  hpe_val
)
}
\arguments{
\item{sync_tags}{Table of data of sync tags}

\item{hpe_col_name}{String name of column in sync_tags that has HPE values}

\item{hpem_col_name}{String name of column in sync_tags that has HPEm values}

\item{id_col_name}{String name of column that includes tag ID (needs to be the same across data tables)}

\item{hpe_val}{Integer value for max HPE value user wants to display the graph to (with the max being the largest HPE value)}
}
\value{
Output of plot HPE versus measured error to the median point with white circles with black outline and red x representing the two dimensional root mean square error of x and y components of error within an HPE bin
}
\description{
Calculate linear regression given HPE level or percentage of data.
Accepts only one of hpem_max or percentage.
}
\examples{
# Load dataset
sample_file <- system.file("extdata", "dummy_sync.csv", package = "filterTelemetry")
dummy_sync <- read.table(file = sample_file, header=TRUE, sep=",")
# Example
sync_linreg <- hpe_linear_regression(dummy_sync, "HPE", "HPEm", "Id", 30)

}
