% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.eqn.R
\name{lm.eqn}
\alias{lm.eqn}
\title{Function to put output of linear model into a usable text format -
this works if model object is simple / for 1 model}
\usage{
lm.eqn(model)
}
\arguments{
\item{model}{Output of lm()}
}
\value{
Character
}
\description{
Function to put output of linear model into a usable text format -
this works if model object is simple / for 1 model
}
\examples{
# Load dataset
dummy_sync <- read.table(file = './extdata/dummy_sync.csv', header=TRUE, sep=",")
# Create a linear model
mod <- lm(dummy_sync[, "HPEm"]~cut_sync_tags[, "HPE"])
# Example use of function
title <- lm.eqn(mod)

}
