% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_eqn.R
\name{lm_eqn}
\alias{lm_eqn}
\title{Function to put output of linear model into a usable text format -
this works if model object is simple / for 1 model}
\usage{
lm_eqn(model)
}
\arguments{
\item{model}{Output of lm()}
}
\value{
Character
}
\description{
Function to put output of linear model into a usable text format -
this works if model object is simple / for 1 model
}
