% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop.filter.R
\name{loop.filter}
\alias{loop.filter}
\title{Subset tags from one a table}
\usage{
loop.filter(data.from, loop.vec, column.name)
}
\arguments{
\item{data.from}{Table that user wants to sort through}

\item{loop.vec}{Vector of values that user wants to find}

\item{column.name}{Name of column that contains the values within loop.vec}
}
\value{
A dataframe with all rows from original table whose specified column includes a value in loop.vec
}
\description{
Subset tags from one a table
}
\examples{
data <- loop.filter(tags, tagnames.vec, id_col)

}
