% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_obs_per_tag.R
\name{num_obs_per_tag}
\alias{num_obs_per_tag}
\title{Get number of observations per tag when given all data or data is sorted
by species, sex, grouping, or a combination, can output graph. Grouping can
be used for a variety of different separations of data such as treatment.}
\usage{
num_obs_per_tag(
  merged_tags,
  id_col_name,
  graph = FALSE,
  scatterplot = FALSE,
  metadata = NULL,
  grouping = NULL,
  grouping_col_name = NULL,
  continuous_grouping = NULL,
  continuous_col_name = NULL,
  size_col_name = NULL,
  start_date = NULL,
  end_date = NULL,
  time_interval = NULL,
  time_col_name = NULL
)
}
\arguments{
\item{merged_tags}{Table of tag data}

\item{id_col_name}{String name of column that has tag IDs}

\item{graph}{Boolean, whether or not user wants to output a bar graph of number of observations per tag, default to FALSE}

\item{scatterplot}{Boolean, whether or not user wants to output a scatterplot of number of observations per size, default to FALSE}

\item{metadata}{Table of data that has metadata for tagged animals, must have column of ID tags}

\item{grouping}{Vector of strings, grouping name, default to NULL}

\item{grouping_col_name}{Vector of strings, name of column that includes grouping, default to NULL, required if grouping!=NULL}

\item{continuous_grouping}{String, name of grouping of continuous variable, default to NULL}

\item{continuous_col_name}{String, name of column that includes continuous grouping, default to NULL, required if grouping!=NULL}

\item{size_col_name}{String, name of column that includes size of animal, default to FALSE, required if scatterplot=TRUE}

\item{start_date}{String, start date in format YYYY-MM-DD, default to NULL}

\item{end_date}{String, end date in format YYYY-MM-DD, default to NULL}

\item{time_interval}{String, interval of time to separate data into, either "week" or "month", default to NULL}

\item{time_col_name}{String, name of column that includes date/time data, default to NULL}
}
\value{
Vector of number of total number of observations and total number of tags (for given parameters) and graph if graph=TRUE or scatterplot=TRUE,
or returns integers of number of observations per week/month if date range given
}
\description{
Get number of observations per tag when given all data or data is sorted
by species, sex, grouping, or a combination, can output graph. Grouping can
be used for a variety of different separations of data such as treatment.
}
\examples{
# Load data
sample_file <- system.file("extdata", "dummy_animals.csv", package = "filterTelemetry")
animal_tag_data <- read.table(file = sample_file, header=TRUE, sep=",")
sample_metadata <- system.file("extdata", "dummy_metadata.csv", package = "filterTelemetry")
metadata <- read.table(file = sample_metadata, header=TRUE, sep=",")
animal_merged <- merge(animal_tag_data, metadata, by = "Id")

# Example to make a scatterplot
num_obs_per_tag_size <- num_obs_per_tag(animal_merged, "Id", graph=TRUE,
                                        grouping=c("Cancer irroratus", "M"),
                                        grouping_col_name=c("Sp", "Sex"),
                                        continuous_grouping=c("Size"),
                                        continuous_col_name=c("Size"),
                                        metadata=metadata, scatterplot=TRUE)

}
