% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prelim_report.R
\name{prelim_report}
\alias{prelim_report}
\title{Wrapper function to generate preliminary report.}
\usage{
prelim_report(
  Tags,
  SyncTags = NULL,
  Metadata,
  ani_tags,
  sync_tags,
  ref_tags,
  id_col_name,
  vec_species,
  hpe_col_name,
  hpem_col_name,
  bins = NULL,
  num_bins,
  lat_col_name,
  long_col_name,
  sp_col_name,
  output_file
)
}
\arguments{
\item{Tags}{File pathway to table with tags, needs to be comma delineated}

\item{SyncTags}{File pathway to table with sync tags (if separate from file with tags), needs to be comma delineated}

\item{Metadata}{File pathway to table of metadata for all animal tags, needs to be semicolon delineated}

\item{ani_tags}{Vector of all animal tags (e.g. c(15820, 14325, 86402))}

\item{sync_tags}{Vector of all sync tags (e.g. c(65106))}

\item{ref_tags}{Vector of all reference tags (e.g. c(65321, 65322))}

\item{id_col_name}{String name of column that includes tag ID (needs to be the same across data tables)}

\item{vec_species}{Vector of string species names (e.g. c("Cancer irroratus","Homarus americanus"))}

\item{hpe_col_name}{String name of column that includes HPE (needs to be the same for all tag data tables)}

\item{hpem_col_name}{String name of column that includes HPEm}

\item{bins}{Vector of doubles for HPE separations to make bins}

\item{num_bins}{Double number of bins that user wants HPE separated into}

\item{lat_col_name}{String name of column that includes latitude}

\item{long_col_name}{String name of column that includes longitude}

\item{sp_col_name}{String name of column that includes species name in metadata file}

\item{output_file}{String for output file name/location}
}
\description{
Wrapper function to generate preliminary report.
}
