% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/receiver_detections.R
\name{receiver_detections}
\alias{receiver_detections}
\title{Function that outputs a scatterplot of all observations of sync tags
with longitude as the x-axis and latitude as the y-axis and the colour of
each point represents HPE.}
\usage{
receiver_detections(
  sync_tags = NULL,
  sync_hpe_col_name = NULL,
  animal_tags = NULL,
  ani_hpe_col_name = NULL,
  lat_col_name,
  long_col_name,
  bins = NULL,
  id_col_name
)
}
\arguments{
\item{sync_tags}{Table of data of sync tags}

\item{sync_hpe_col_name}{String name of column in sync_tags that has HPE values}

\item{animal_tags}{Table of data of animal tags}

\item{ani_hpe_col_name}{String name of column in animal_tags that has HPE values}

\item{lat_col_name}{String name of column in animal_tags that has latitude}

\item{long_col_name}{String name of column in animal_tags that has longitude}

\item{bins}{Vector of doubles of bins that user wants HPE to be split into, max vector length 6. If empty, defaults to c(4,6,10,50,250)}

\item{id_col_name}{String name of column in sync_tags that has tag ID}
}
\description{
Function that outputs a scatterplot of all observations of sync tags
with longitude as the x-axis and latitude as the y-axis and the colour of
each point represents HPE.
}
\examples{
# Load data
sample_file <- system.file("extdata", "dummy_sync.csv", package = "filterTelemetry")
sync_tag_data <- read.table(file = sample_file, header=TRUE, sep=",")
# Example
receiver_detections(sync_tags=sync_tag_data, sync_hpe_col_name="HPE", bins=7,
                    lat_col_name="Latitude", long_col_name="Longitude", id_col_name="Id")

}
