% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removed.points.R
\name{removed.points}
\alias{removed.points}
\title{A summary of what points are removed given the current filtering strategy.}
\usage{
removed.points(
  sync_tags,
  hpe_col_name,
  hpem_col_name,
  id_col_name,
  time_col_name,
  percentage = NULL,
  hpem_max = NULL,
  df_return = FALSE
)
}
\arguments{
\item{sync_tags}{Table of data of sync tags}

\item{hpe_col_name}{String name of column in sync_tags that has HPE values}

\item{hpem_col_name}{String name of column in sync_tags that has HPEm values}

\item{percentage}{Percentage of data you are willing to lose, default to NULL}

\item{hpem_max}{Maximum HPEm value you are willing to have the data be, default to NULL}

\item{df_return}{Default to FALSE, if TRUE then returns dataframe of all removed points. Otherwise, check return statement.}
}
\value{
If df_return is FALSE, returns list of three return values (total number of removed points, number of removed points by tag, number of removed points by week).
If df_return is TRUE, returns dataframe of all removed points.
}
\description{
A summary of what points are removed given the current filtering strategy.
}
\examples{
# Load data
sync_tag_data <- read.table(file = './extdata/dummy_sync.csv', header=TRUE, sep=",")
# Example
sync_rempoint <- removed.points(sync_tags=sync_tag_data, hpe_col_name="HPE",
  hpem_col_name="HPEm", id_col_name="Id",
  time_col_name="Time", percentage=0.1,
  df_return=FALSE)

}
