% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removed_points.R
\name{removed_points}
\alias{removed_points}
\title{A summary of what points are removed given the current filtering strategy.}
\usage{
removed_points(
  sync_tags,
  hpe_col_name,
  hpem_col_name,
  id_col_name,
  time_col_name,
  date_format = "\%Y/\%m/\%d \%H:\%M:\%OS",
  percentage = NULL,
  hpem_max = NULL,
  df_return = FALSE
)
}
\arguments{
\item{sync_tags}{Table of data of sync tags}

\item{hpe_col_name}{String name of column in sync_tags that has HPE values}

\item{hpem_col_name}{String name of column in sync_tags that has HPEm values}

\item{id_col_name}{String name of column in sync_tags that has tag ID}

\item{time_col_name}{String, name of column that includes date/time data, default to NULL}

\item{date_format}{Character string giving a date-time format as used in base as.POSIX, default to "\%Y/\%m/\%d \%H:\%M:\%OS"}

\item{percentage}{Percentage of data you are willing to lose, default to NULL}

\item{hpem_max}{Maximum HPEm value you are willing to have the data be, default to NULL}

\item{df_return}{Default to FALSE, if TRUE then returns dataframe of all removed points. Otherwise, check return statement.}
}
\value{
If df_return is FALSE, returns list of three return values (total number of removed points, number of removed points by tag, number of removed points by week).
If df_return is TRUE, returns dataframe of all removed points.
}
\description{
A summary of what points are removed given the current filtering strategy.
}
\examples{
# Load data
sample_file <- system.file("extdata", "dummy_sync.csv", package = "filterTelemetry")
sync_tag_data <- read.table(file = sample_file, header=TRUE, sep=",")
# Example
sync_rempoint <- removed_points(sync_tags=sync_tag_data, hpe_col_name="HPE",
  hpem_col_name="HPEm", id_col_name="Id",
  time_col_name="Time", percentage=0.1,
  df_return=FALSE)

}
