% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/second.report.R
\name{second.report}
\alias{second.report}
\title{Wrapper function to generate secondary report that outlines how the data
will appear after certain processing with HPE and data loss percentages.}
\usage{
second.report()
}
\description{
Wrapper function to generate secondary report that outlines how the data
will appear after certain processing with HPE and data loss percentages.
}
