% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/second_report.R
\name{second_report}
\alias{second_report}
\title{Wrapper function to generate secondary report that outlines how the data
will appear after certain processing with HPE and HPEm.}
\usage{
second_report(
  sync_tag_data,
  hpe_col_name,
  hpem_col_name,
  id_col_name,
  time_col_name,
  date_format = "\%Y/\%m/\%d \%H:\%M:\%OS",
  hpe_val,
  hpem_max,
  output_file
)
}
\arguments{
\item{sync_tag_data}{Table of data of sync tags}

\item{hpe_col_name}{String name of column that includes HPE (needs to be the same for all tag data tables)}

\item{hpem_col_name}{String name of column that includes HPEm}

\item{id_col_name}{String name of column that includes tag ID (needs to be the same across data tables)}

\item{time_col_name}{String, name of column that includes date/time data}

\item{date_format}{Character string giving a date-time format as used in base as.POSIX, default to "\%Y/\%m/\%d \%H:\%M:\%OS"}

\item{hpe_val}{Integer value for max HPE value user wants to display the graph to (with the max being the largest HPE value)}

\item{hpem_max}{Maximum HPEm value you are willing to have the data be}

\item{output_file}{String for output file name/location}
}
\description{
Wrapper function to generate secondary report that outlines how the data
will appear after certain processing with HPE and HPEm.
}
