% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_tag_data.R
\name{split_tag_data}
\alias{split_tag_data}
\title{Subset tags from files}
\arguments{
\item{tags}{Table of all tags or just animal tags (if there is a sync tag file)}

\item{sync_tags}{Table of all sync tags, default to NULL}

\item{id_col}{String that is what the column for id/transmitter is in all tables}

\item{sync_tags.vec}{Vector of strings, tag numbers for all sync tags}

\item{ref_tags.vec}{Vector of strings, tag numbers for all reference tags}

\item{ani_tags.vec}{Vector of strings, tag numbers for all animal tags}
}
\value{
A list of three dataframes that give animal tag data, sync tag data, and reference tag data
}
\description{
Subset tags from files
}
\examples{
# Load data
sample_file_animals <- system.file("extdata", "dummy_animals.csv", package = "filterTelemetry")
sample_file_sync <- system.file("extdata", "dummy_sync.csv", package = "filterTelemetry")
all_tags <- read.table(file = sample_file_animals, header=TRUE, sep=",")
other_tags <- read.table(file = sample_file_sync, header=TRUE, sep=",")
ani_tags <- unique(all_tags$Id)
non_ref <- subset(other_tags, Id!=61034)
sync_tags <- unique(non_ref$Id)
ref_tags <- c(61034)
# Example
test_ <- split_tag_data(tags = all_tags, id_col = "Id",
                     sync_tags.vec = sync_tags, ref_tags.vec = ref_tags,
                     ani_tags.vec = ani_tags)

}
