% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.summary.R
\name{sync.summary}
\alias{sync.summary}
\title{Get summary statistics for sync tags at chosen HPE values.
Summary statistics includes mean, maximum, minimum, standard deviation, and standard error.}
\usage{
sync.summary(sync_tags, hpe_col_name, hpe_levels)
}
\arguments{
\item{sync_tags}{Table of data of sync tags}

\item{hpe_col_name}{String name of column in sync_tags that has HPE values}

\item{hpe_levels}{Vector of HPE values}
}
\value{
Dataframe of numerics of summary statistics for sync data
}
\description{
Get summary statistics for sync tags at chosen HPE values.
Summary statistics includes mean, maximum, minimum, standard deviation, and standard error.
}
