% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbinned_hpe_quantiles.R
\name{unbinned.hpe.quantiles}
\alias{unbinned.hpe.quantiles}
\title{Get HPE quantiles corresponding to the given or default probabilities
and output scatterplots that showed HPE values vs HPEm.}
\usage{
unbinned.hpe.quantiles(
  sync_tags,
  hpe_col_name,
  hpem_col_name,
  quantiles = NULL
)
}
\arguments{
\item{sync_tags}{Table of data of sync tags}

\item{hpe_col_name}{String name of column in sync_tags that has HPE values}

\item{hpem_col_name}{String name of column in sync_tags that has HPEm values}

\item{quantiles}{Vector of numerics, maximum 6 doubles, default to NULL. When NULL, quantiles = c(0.95, 0.9, 0.75, 0.5, 0.25)}
}
\value{
Dataframe of HPE values to cut off at for each quantile.
}
\description{
Get HPE quantiles corresponding to the given or default probabilities
and output scatterplots that showed HPE values vs HPEm.
}
\examples{
# load dataset:
sample_file <- system.file("extdata", "dummy_sync.csv", package = "filterTelemetry")
sync_tag_data <- read.table(file = sample_file, header=TRUE, sep=",")
quantiles <- c(0.9, 0.8, 0.5, 0.3, 0.1, 0)
sync_hpe2 <- unbinned.hpe.quantiles(sync_tag_data, "HPE", "HPEm", quantiles)

}
