% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpe.linear.regression.R
\name{hpe.linear.regression}
\alias{hpe.linear.regression}
\title{Calculate linear regression given HPE level or percentage of data.
Accepts only one of hpem_max or percentage.}
\arguments{
\item{sync_tags}{Table of data of sync tags}

\item{hpe_col_name}{String name of column in sync_tags that has HPE values}

\item{hpem_col_name}{String name of column in sync_tags that has HPEm values}

\item{percentage}{Percentage of data you are willing to lose, default to NULL}

\item{hpem_max}{Maximum HPEm value you are willing to have the data be, default to NULL}
}
\value{
Vector of three values (p-value, r-squared, and percentage of removed data as a decimal).
}
\description{
Calculate linear regression given HPE level or percentage of data.
Accepts only one of hpem_max or percentage.
}
\examples{

sync_tag_data <- read.table(file = "/inst/extdata/dummy_sync.csv", header=TRUE, sep=",")
sync_linreg <- hpe.linear.regression(sync_tag_data, "HPE", "HPEm", percentage=0.1)
}
