% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num.obs.per.tag.R
\name{num.obs.per.tag}
\alias{num.obs.per.tag}
\title{Get number of observations per tag when given all data or data is sorted
by species or sex, can output graph}
\usage{
num.obs.per.tag(
  merged_tags,
  id_col_name,
  graph = FALSE,
  sp = NULL,
  sp_col_name = NULL,
  sex = NULL,
  sex_col_name = NULL
)
}
\arguments{
\item{merged_tags}{Table of data (requires metadata if using sp or sex)}

\item{id_col_name}{String name of column that has tag IDs}

\item{graph}{Boolean, whether or not user wants an outputted bar graph of number of observations per tag, default to FALSE}

\item{sp}{String, species name, default to FALSE}

\item{sp_col_name}{String, name of column that includes species name, default to FALSE, required if sp!=NULL}

\item{sex}{String, animal's sex, default to FALSE}

\item{sex_col_name}{String, name of column that includes animal's sex, default to FALSE, required if sex!=NULL}
}
\value{
Vector of number of total number of observations and total number of tags (for given parameters) and graph if graph=TRUE
}
\description{
Get number of observations per tag when given all data or data is sorted
by species or sex, can output graph
}
\examples{
num_obs_per_tag_lob <- num.obs.per.tag(animal_merged, "Id", graph=TRUE,
sp="Homarus americanus", sp_col_name="Sp")

}
