% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num.obs.per.tag.R
\name{num.obs.per.tag}
\alias{num.obs.per.tag}
\title{Get number of observations per tag when given all data or data is sorted
by species, sex, grouping, or a combination, can output graph. Grouping can
be used for a variety of different separations of data such as treatment.}
\usage{
num.obs.per.tag(
  merged_tags,
  id_col_name,
  graph = FALSE,
  scatterplot = FALSE,
  metadata = NULL,
  sp = NULL,
  sp_col_name = NULL,
  sex = NULL,
  sex_col_name = NULL,
  grouping = NULL,
  grouping_col_name = NULL,
  size_col_name = NULL,
  start_date = NULL,
  end_date = NULL,
  time_interval = NULL,
  time_col_name = NULL
)
}
\arguments{
\item{merged_tags}{Table of tag data}

\item{id_col_name}{String name of column that has tag IDs}

\item{graph}{Boolean, whether or not user wants to output a bar graph of number of observations per tag, default to FALSE}

\item{scatterplot}{Boolean, whether or not user wants to output a scatterplot of number of observations per size, default to FALSE}

\item{metadata}{Table of data that has metadata for tagged animals, must have column of ID tags}

\item{sp}{String, species name, default to FALSE}

\item{sp_col_name}{String, name of column that includes species name, default to FALSE, required if sp!=NULL}

\item{sex}{String, animal's sex, default to FALSE}

\item{sex_col_name}{String, name of column that includes animal's sex, default to FALSE, required if sex!=NULL}

\item{grouping}{String, grouping name, default to FALSE}

\item{grouping_col_name}{String, name of column that includes grouping, default to FALSE, required if grouping!=NULL}

\item{size_col_name}{String, name of column that includes size of animal, default to FALSE, required if scatterplot=TRUE}

\item{start_date}{String, start date in format YYYY-MM-DD, default to NULL}

\item{end_date}{String, end date in format YYYY-MM-DD, default to NULL}

\item{time_interval}{String, interval of time to separate data into, either "week" or "month", default to NULL}

\item{time_col_name}{String, name of column that includes date/time data, default to NULL}
}
\value{
Vector of number of total number of observations and total number of tags (for given parameters) and graph if graph=TRUE or scatterplot=TRUE,
or returns integers of number of observations per week/month if date range given
}
\description{
Get number of observations per tag when given all data or data is sorted
by species, sex, grouping, or a combination, can output graph. Grouping can
be used for a variety of different separations of data such as treatment.
}
\examples{
num_obs_per_tag_lob <- num.obs.per.tag(animal_merged, "Id", graph=TRUE,
sp="Homarus americanus", sp_col_name="Sp")

}
